+++ event-sourcing/event-publisher/event-publisher-timer/src/main/java/uk/gov/justice/services/eventsourcing/publishedevent/publishing/PublisherTimerBean.java
@@ -1,6 +1,7 @@
 package uk.gov.justice.services.eventsourcing.publishedevent.publishing;
 
 import uk.gov.justice.services.eventsourcing.util.jee.timer.TimerServiceManager;
+import uk.gov.justice.services.jmx.logging.MdcLoggerInterceptor;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.Resource;
@@ -9,9 +10,11 @@ import javax.ejb.Startup;
 import javax.ejb.Timeout;
 import javax.ejb.TimerService;
 import javax.inject.Inject;
+import javax.interceptor.Interceptors;
 
 @Singleton
 @Startup
+@Interceptors(MdcLoggerInterceptor.class)
 public class PublisherTimerBean {
 
     private static final String TIMER_JOB_NAME = "event-store.de-queue-events-and-publish.job";